fix/refactor code
