+++ .eslintrc.json
@@ -24,7 +24,7 @@
   "rules": {
     "@typescript-eslint/indent": ["error", 2],
     "import/prefer-default-export": "off",
-    "@typescript-eslint/no-explicit-any": "error",
+    "@typescript-eslint/no-explicit-any": "off",
     "@typescript-eslint/no-unused-vars": "error",
     "@typescript-eslint/no-useless-constructor": "error",
     "react/function-component-definition": [
