+++ index.txt
@@ -1,0 +1,3 @@
+yarn run v1.22.19
+$ jest
+info Visit https://yarnpkg.com/en/docs/cli/run for documentation about this command.
