--- src/components/sidebar/SidebarMenu.scss
@@ -1,194 +1,0 @@
-@mixin transition($property) {
-  transition: $property 0.2 ease-in-out;
-}
-
-.datagrid-container {
-  display: flex;
-  background: #f2f2f2;
-
-  .datagrid-container__grid {
-    background: #f2f2f2;
-    padding: 24px;
-  }
-}
-
-.sidebar-menu {
-  background: #233043;
-  color: #8d91a0;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  border-top-right-radius: 8px;
-  border-bottom-right-radius: 8px;
-
-  @include transition(width);
-
-  &.expanded {
-    width: 300px;
-  }
-
-  &.collapsed {
-    width: 90px;
-  }
-}
-
-.sidebar-line {
-  border: 0.031rem solid #8d91a0;
-  margin-top: 30px;
-  background: #8d91a0;
-}
-
-.sidebar_menu-button {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 30px;
-  margin-left: 25px;
-  margin-top: 24px;
-  .sidebar_span {
-    color: #fff;
-  }
-  .sidebar_menu-button-icon {
-    cursor: pointer;
-    &.hover {
-      color: white;
-    }
-  }
-}
-
-.search-content-collapsed {
-  display: flex;
-  justify-content: center;
-  margin-top: 10px;
-  margin-right: 30px;
-}
-
-.search-content-expanded {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  margin-top: 10px;
-  background: #f6f8fa;
-  border-radius: 5.54012px;
-  width: 278px;
-  height: 46px;
-  align-items: center;
-  margin-left: 10px;
-}
-
-.sidebar_search-input {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  overflow: hidden;
-  white-space: nowrap;
-  padding-left: 20px;
-}
-
-.sidebar_search-icon {
-  padding-left: 0px;
-}
-
-.sidebar_input-search {
-  background: #f6f8fa;
-  border: none;
-  outline: none;
-  color: #8d91a0;
-  padding-left: 10px;
-  display: flex;
-}
-
-.sidebar_menu-item-view {
-  text-align: center;
-  display: flex;
-  align-items: center;
-  margin: 0 5px;
-
-  a {
-    color: #8d91a0;
-    text-decoration: none;
-    display: block;
-    padding: 10px;
-    width: 100%;
-    border-radius: 5px;
-  }
-}
-
-.item-content {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  overflow: hidden;
-  width: 100%;
-  white-space: nowrap;
-  @include transition(width);
-
-  &.collapsed {
-    width: 35px;
-  }
-
-  .icon {
-    width: 20px;
-    height: 24px;
-    margin-left: 10px;
-  }
-
-  &.hover {
-    & + .tooltipe {
-      display: inline-block;
-      animation-duration: 0.4s;
-      animation-name: tooltipe;
-    }
-  }
-}
-
-.sidebar_label {
-  font-weight: 500px;
-  margin-left: 30px;
-  overflow: hidden;
-}
-
-.sidebar_menu-card-view {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  text-align: center;
-  gap: 15px;
-  &.collapsed {
-    justify-content: center;
-  }
-
-  .sidebar_profile-img {
-    border-radius: 50px;
-    box-sizing: border-box;
-    margin-left: 20px;
-    width: 40px;
-    height: 40px;
-    &.collapsed {
-      width: 30px;
-      height: 30px;
-      margin-left: 0;
-      margin-right: 20px;
-    }
-  }
-  .sidebar_profile-info {
-    display: flex;
-    align-items: center;
-    overflow: hidden;
-    height: 80px;
-    @include transition(height);
-
-    &.collapsed {
-      opacity: 0;
-      height: 0;
-    }
-  }
-  .sidebar_name {
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 25px;
-    letter-spacing: -0.01em;
-    color: #8d91a0;
-  }
-}
