+++ src/components/sidebar/SidebarMenuItemView/SidebarMenuItemView.scss
@@ -1,0 +1,46 @@
+@mixin transition($property) {
+  transition: $property 0.2 ease-in-out;
+}
+
+.sidebar_menu-item-view {
+  text-align: center;
+  display: flex;
+  align-items: center;
+  margin: 0 5px;
+
+  a {
+    display: flex !important;
+    color: #8d91a0;
+    text-decoration: none;
+    padding: 10px;
+    width: 100%;
+    border-radius: 5px;
+  }
+  a {
+    display: flex;
+
+    &:hover {
+      color: #535bf2;
+    }
+  }
+  .item-content {
+    width: 100%;
+    @include transition(width);
+
+    .sidebar_label {
+      font-weight: 500px;
+      margin-left: 30px;
+      overflow: hidden;
+    }
+
+    &.collapsed {
+      width: 35px;
+    }
+
+    .icon {
+      width: 20px;
+      height: 24px;
+      margin-left: 10px;
+    }
+  }
+}
