+++ src/components/sidebar/SidebarPage.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { ISidebarMenuCard } from '../../interfaces/index';
-import { SidebarMenu } from './SidebarMenu';
+import { SidebarMenu } from './SidebarMenu/SidebarMenu';
 import imageProfile from '../../assets/img_cv.jpg';
-import './SidebarMenu.scss';
 import { items } from '../../routes';
 
 export const SidebarPage = () => {
@@ -12,9 +11,5 @@ export const SidebarPage = () => {
     photoUrl: imageProfile,
   };
 
-  return (
-    <div className="sidebar-page">
-      <SidebarMenu items={items} card={card} />
-    </div>
-  );
+  return <SidebarMenu items={items} card={card} />;
 };
